refactor: modularize websocket and irc client
